requirements.txt
openai>=1.0.0
anthropic>=0.5.0
langchain>=0.0.267
llama-index>=0.8.0
streamlit>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
jinja2>=3.1.2
